requirements.txt
streamlit
pytesseract
PyMuPDF
Pillow
opencv-python-headless
numpy
pandas
